Integrate base/format/osl changes into spiral-core

+++ core/src/jvmMain/kotlin/info/spiralframework/core/formats/archives/ZipFormat.kt
@@ -1,16 +1,22 @@
 package info.spiralframework.core.formats.archives
 
 import info.spiralframework.base.common.SpiralContext
-import info.spiralframework.base.util.copyFromStream
-import info.spiralframework.base.util.copyToStream
+import info.spiralframework.base.common.io.FlowOutputStream
+import info.spiralframework.base.common.io.readChunked
 import info.spiralframework.core.formats.*
-import info.spiralframework.formats.archives.*
-import info.spiralframework.formats.archives.srd.SRDEntry
-import info.spiralframework.formats.utils.DataSource
-import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import java.io.OutputStream
+import info.spiralframework.formats.common.archives.*
+import info.spiralframework.formats.common.archives.srd.BaseSrdEntry
+import info.spiralframework.formats.common.archives.srd.SrdArchive
+import kotlinx.coroutines.*
+import kotlinx.coroutines.channels.Channel
+import org.abimon.kornea.io.common.DataSource
+import org.abimon.kornea.io.common.addCloseHandler
+import org.abimon.kornea.io.common.copyToOutputFlow
+import org.abimon.kornea.io.common.flow.OutputFlow
+import org.abimon.kornea.io.common.use
+import org.abimon.kornea.io.jvm.JVMOutputFlow
+import org.abimon.kornea.io.jvm.files.FileDataSource
+import java.io.*
 import java.util.*
 import java.util.zip.ZipEntry
 import java.util.zip.ZipFile
@@ -30,32 +36,43 @@ object ZipFormat : ReadableSpiralFormat<ZipFile>, WritableSpiralFormat {
      *
      * @return a FormatResult containing either [T] or null, if the stream does not contain the data to form an object of type [T]
      */
-    override fun read(context: SpiralContext, readContext: FormatReadContext?, source: DataSource): FormatResult<ZipFile> {
-        source().use { stream ->
-            //TODO: Switch over to data sources
-//            val possibleFile = stream.path?.let(::File)
-//            if (possibleFile?.exists() == true) {
-//                try {
-//                    return FormatResult.Success(this, ZipFile(possibleFile), 1.0)
-//                } catch (io: IOException) {
-//                    return FormatResult.Fail(this, 1.0, io)
-//                }
-//            } else {
-                val zip: ZipFile
-                val tmpFile = File.createTempFile(UUID.randomUUID().toString(), ".dat")
-                tmpFile.deleteOnExit()
+    override suspend fun read(context: SpiralContext, readContext: FormatReadContext?, source: DataSource<*>): FormatResult<ZipFile> {
+        if (source is FileDataSource) {
+            try {
+                return withContext(Dispatchers.IO) {
+                    FormatResult.Success(this@ZipFormat, ZipFile(source.backing), 1.0)
+                }
+            } catch (io: IOException) {
+                return FormatResult.Fail(this, 1.0, io)
+            }
+        } else {
+            var zip: ZipFile? = null
+            var ioException: IOException? = null
+            val tmpFile = withContext(Dispatchers.IO) { File.createTempFile(UUID.randomUUID().toString(), ".dat") }
+            tmpFile.deleteOnExit()
 
+            withContext(Dispatchers.IO) {
                 try {
-                    FileOutputStream(tmpFile).use(stream::copyToStream)
-                    zip = ZipFile(tmpFile)
+                    source.openInputFlow()?.use { flow ->
+                        withContext(Dispatchers.IO) {
+                            JVMOutputFlow(FileOutputStream(tmpFile)).use(flow::copyToOutputFlow)
+                            zip = ZipFile(tmpFile)
+                        }
+
+                        source.addCloseHandler { tmpFile.delete() }
+                    }
                 } catch (io: IOException) {
                     tmpFile.delete()
-
-                    return FormatResult.Fail(this, 1.0, io)
+                    ioException = io
                 }
+            }
 
-                return FormatResult.Success(this, zip, 1.0)
-//            }
+            if (zip != null) {
+                return FormatResult.Success(this, zip!!, 1.0)
+            } else {
+                tmpFile.delete()
+                return FormatResult.Fail(this, 1.0, ioException)
+            }
         }
     }
 
@@ -68,7 +85,7 @@ object ZipFormat : ReadableSpiralFormat<ZipFile>, WritableSpiralFormat {
      *
      * @return If we are able to write [data] as this format
      */
-    override fun supportsWriting(context: SpiralContext, data: Any): Boolean = data is IArchive || data is ZipFile
+    override fun supportsWriting(context: SpiralContext, writeContext: FormatWriteContext?, data: Any): Boolean = data is AwbArchive || data is WadArchive || data is CpkArchive || data is SpcArchive || data is PakArchive || data is ZipFile
 
     /**
      * Writes [data] to [stream] in this format
@@ -81,50 +98,54 @@ object ZipFormat : ReadableSpiralFormat<ZipFile>, WritableSpiralFormat {
      *
      * @return An enum for the success of the operation
      */
-    override fun write(context: SpiralContext, writeContext: FormatWriteContext?, data: Any, stream: OutputStream): FormatWriteResponse {
-        val zipOut = ZipOutputStream(stream)
+    override suspend fun write(context: SpiralContext, writeContext: FormatWriteContext?, data: Any, flow: OutputFlow): FormatWriteResponse {
+        val zipOut = ZipOutputStream(FlowOutputStream.withGlobalScope(flow, false))
 
-        try {
-            when (data) {
-                is ZipFile -> data.entries().iterator().forEach { entry ->
-                    zipOut.putNextEntry(entry)
-                    data.getInputStream(entry).use(zipOut::copyFromStream)
-                }
-
-                is AWB -> data.entries.forEach { entry ->
-                    zipOut.putNextEntry(ZipEntry(entry.id.toString()))
-                    entry.inputStream.use(zipOut::copyFromStream)
-                }
-                is CPK -> data.files.forEach { entry ->
-                    zipOut.putNextEntry(ZipEntry(entry.name))
-                    entry.inputStream.use(zipOut::copyFromStream)
-                }
-                is Pak -> data.files.forEach { entry ->
-                    zipOut.putNextEntry(ZipEntry(entry.index.toString()))
-                    entry.inputStream.use(zipOut::copyFromStream)
-                }
-                is SPC -> data.files.forEach { entry ->
-                    zipOut.putNextEntry(ZipEntry(entry.name))
-                    entry.inputStream.use(zipOut::copyFromStream)
-                }
-                is SRD -> data.entries.groupBy(SRDEntry::dataType).forEach { (_, list) ->
-                    list.forEachIndexed { index, entry ->
-                        zipOut.putNextEntry(ZipEntry("${entry.dataType}-$index-data"))
-                        entry.dataStream.use(zipOut::copyFromStream)
-                        zipOut.putNextEntry(ZipEntry("${entry.dataType}-$index-subdata"))
-                        entry.subdataStream.use(zipOut::copyFromStream)
+        return withContext(Dispatchers.IO) {
+            try {
+                when (data) {
+                    is ZipFile -> data.entries().iterator().forEach { entry ->
+                        zipOut.putNextEntry(entry)
+                        data.getInputStream(entry).use { zipIn -> zipIn.copyTo(zipOut) }
                     }
-                }
-                is WAD -> data.files.forEach { entry ->
-                    zipOut.putNextEntry(ZipEntry(entry.name))
-                    entry.inputStream.use(zipOut::copyFromStream)
-                }
-                else -> return FormatWriteResponse.WRONG_FORMAT
-            }
-        } finally {
-            zipOut.finish()
-        }
 
-        return FormatWriteResponse.SUCCESS
+                    is AwbArchive -> data.files.forEach { entry ->
+                        zipOut.putNextEntry(ZipEntry(entry.id.toString()))
+                        data.openFlow(entry)?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                    }
+                    is CpkArchive -> data.files.forEach { entry ->
+                        zipOut.putNextEntry(ZipEntry(entry.name))
+                        data.openDecompressedFlow(context, entry)
+                                ?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                    }
+                    is PakArchive -> data.files.forEach { entry ->
+                        zipOut.putNextEntry(ZipEntry(entry.index.toString()))
+                        data.openFlow(entry)?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                    }
+                    is SpcArchive -> data.files.forEach { entry ->
+                        zipOut.putNextEntry(ZipEntry(entry.name))
+                        data.openDecompressedFlow(context, entry)
+                                ?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                    }
+                    is SrdArchive -> data.entries.groupBy(BaseSrdEntry::classifierAsString).forEach { (_, list) ->
+                        list.forEachIndexed { index, entry ->
+                            zipOut.putNextEntry(ZipEntry("${entry.classifierAsString}-$index-data"))
+                            entry.openMainDataFlow()?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                            zipOut.putNextEntry(ZipEntry("${entry.classifierAsString}-$index-subdata"))
+                            entry.openSubDataFlow()?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                        }
+                    }
+                    is WadArchive -> data.files.forEach { entry ->
+                        zipOut.putNextEntry(ZipEntry(entry.name))
+                        data.openFlow(entry)?.readChunked { buffer, offset, length -> zipOut.write(buffer, offset, length) }
+                    }
+                    else -> return@withContext FormatWriteResponse.WRONG_FORMAT
+                }
+            } finally {
+                zipOut.finish()
+            }
+
+            return@withContext FormatWriteResponse.SUCCESS
+        }
     }
 }
